+++ Xrm.WebApi/Responses/MultipleRecordsResponse.cs
@@ -8,7 +8,7 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace Xrm.WebApi.Response
+namespace Xrm.WebApi.Responses
 {
     /// <summary>
     /// Json model for a Dynamics 365 Xrm Web Api response containing a collection of records.
